Split withMediaPlayer and patch queries
