+++ src/@data/mediaPlayer/albumsQuery.js
@@ -2,7 +2,7 @@ import gql from 'graphql-tag';
 
 export default gql`
   query getAlbums($limit: Int!, $skip: Int!) {
-    content(channel: "newspring_albums", limit: $limit, skip: $skip) {
+    library: content(channel: "newspring_albums", limit: $limit, skip: $skip) {
       id
       entryId: id
       title
