+++ src/@data/mediaPlayer/index.js
@@ -1,5 +1,6 @@
 import { graphql } from 'react-apollo';
 import { compose } from 'recompose';
+import fetchMoreResolver from '@data/utils/fetchMoreResolver';
 import playMutation from './playMutation';
 import pauseMutation from './pauseMutation';
 import mediaPlayerQuery from './mediaPlayerQuery';
@@ -28,23 +29,32 @@ const get = graphql(mediaPlayerQuery, {
   }),
 });
 
-const libraryQuery = graphql(albumsQuery, {
-  props: ({ data: { content } }) => ({
-    library: content,
+export const withLibrary = graphql(albumsQuery, {
+  options: { variables: { limit: 20, skip: 0 } },
+  props: ({ data }) => ({
+    library: data.library,
+    libraryIsLoading: data.loading,
+    refetchLibrary: data.refetch,
+    fetchMore: fetchMoreResolver({
+      collectionName: 'library',
+      data,
+    }),
   }),
 });
 
-const playlistQuery = graphql(albumQuery, {
-  props: ({ data: { content } }) => ({
-    playlist: content,
+export const withPlaylist = graphql(albumQuery, {
+  options: (ownProps = {}) => ({
+    variables: {
+      id: ownProps.id,
+    },
+  }),
+  props: ({ data: { playlist } }) => ({
+    playlist,
   }),
 });
 
-
-export default compose(
+export const withMediaPlayerActions = compose(
   play,
   pause,
   get,
-  libraryQuery,
-  playlistQuery,
 );
